+++ scripts/playground/isEmpty.js
@@ -8,4 +8,4 @@ console.log(isEmpty([]));
 console.log(isEmpty(100));
 console.log(isEmpty(''));
 console.log(isEmpty({}));
-console.log(isEmpty(new ObjectId('5d5110693fb0490f1fc07b5b')));
+console.log(isEmpty(ObjectId('5d5110693fb0490f1fc07b5b')));